refactor(webgl): per-concern split of shader-program.ts
shader-program.ts 1141 → 51 LoC. Extracts the four method groups into
packages/framework/webgl/src/ts/context/shader-program/:

- shader-lifecycle.ts (~274 LoC) — createShader, shaderSource,
  compileShader, getShaderInfoLog, getShaderSource, getShaderParameter,
  deleteShader, isShader
- program-lifecycle.ts (~300 LoC) — createProgram, attachShader,
  detachShader, linkProgram, useProgram, validateProgram,
  getProgramInfoLog, getProgramParameter, deleteProgram, isProgram
- uniforms.ts (~410 LoC) — getUniformLocation + the full uniform{1..4}
  {f,i}{,v} family + uniformMatrix{2,3,4}fv + getUniform
- introspection.ts (~240 LoC) — getAttribLocation, getActiveAttrib,
  getActiveUniform, vertexAttrib* family + parameter getters

Same `install*Methods(proto)` pattern PRs #273/#309 used:
- Typed `<Group>Methods` interface declaration-merged into
  WebGLContextBase via `declare module '../../webgl-context-base.js'`
- Methods object with `ThisType<WebGLContextBase>`
- Per-group `install<Group>Methods(proto)` function called from the
  base file's barrel + composite installer

Base shader-program.ts now re-exports the aggregate `ShaderProgramMethods`
interface (extends all four group interfaces) so existing imports from
context/index.ts keep compiling unchanged. The four per-concern installs
are each imported twice (named + bare side-effect) so tsc preserves the
augmentations in the emitted .d.ts.

context/index.ts's installAllContextMethods aggregator is untouched —
still calls a single installShaderProgramMethods, which now fans out
internally.
